Not possible to connect with already registered nick
Did this by adding a list of nicknames to the server-state, which is updated every time anyone connects/disconnects. Before connecting, the server checks if the users nick is already present on the server. To do that I had to change so that the client also sends his nick when connecting/disconnecting.